+++ WorldEditCommands/WorldEditCommands.cs
@@ -9,7 +9,7 @@ public class WorldEditCommands : BaseUnityPlugin
 {
   public const string GUID = "world_edit_commands";
   public const string NAME = "World Edit Commands";
-  public const string VERSION = "1.46";
+  public const string VERSION = "1.47";
   public void Awake()
   {
     new Harmony(GUID).PatchAll();